replace tools workspace with xtask tasks
