--- tools/change-channel/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-authors = ["Jorge Aparicio <jorge.aparicio@ferrous-systems.com>"]
-edition = "2018"
-license = "MIT OR Apache-2.0"
-name = "change-channel"
-version = "0.0.0"
-
-[dependencies]
-anyhow = "1.0.27"
-consts = { path = "../../advanced/common/consts" }
-hidapi = "1.2.2"
-pids = { path = "../../common/pids" }
