--- tools/dongle-flash/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-authors = ["Jorge Aparicio <jorge.aparicio@ferrous-systems.com>"]
-edition = "2018"
-license = "MIT OR Apache-2.0"
-name = "dongle-flash"
-version = "0.0.0"
-
-[dependencies]
-anyhow = "1.0.31"
-ihex = "1.1.2"
-serialport = "3.3.0"
-tempfile = "3.1.0"
-xmas-elf = "0.7.0"
