--- tools/serial-term/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-authors = ["Jorge Aparicio <jorge.aparicio@ferrous-systems.com>"]
-edition = "2018"
-license = "MIT OR Apache-2.0"
-name = "serial-term"
-version = "0.0.0"
-
-[dependencies]
-anyhow = "1.0.30"
-consts = { path = "../../advanced/common/consts" }
-ctrlc = "3.1.4"
-serialport = "3.3.0"
