--- tools/usb-list/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-authors = ["Jorge Aparicio <jorge.aparicio@ferrous-systems.com>"]
-edition = "2018"
-license = "MIT OR Apache-2.0"
-name = "usb-list"
-version = "0.0.0"
-
-[dependencies]
-consts = { path = "../../advanced/common/consts" }
-pids = { path = "../../common/pids" }
-rusb = "0.5.5"
