+++ xtask/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+authors = ["Jorge Aparicio <jorge.aparicio@ferrous-systems.com>"]
+edition = "2018"
+name = "xtask"
+version = "0.0.0"
+
+[dependencies]
+color-eyre = "0.5.10"
+consts = { path = "../advanced/common/consts" }
+ctrlc = "3.1.4"
+hidapi = "1.2.2"
+ihex = "1.1.2"
+pids = { path = "../common/pids" }
+rusb = "0.5.5"
+serialport = "3.3.0"
+tempfile = "3.2.0"
+xmas-elf = "0.7.0"
