+++ xtask/src/config.rs
@@ -1,0 +1,1 @@
+pub const DONGLE_PATH: &[&str] = &["boards", "dongle"];
